requirement.txt
streamlit==1.16.0
tensorflow==2.17.1
numpy==1.23.5
Pillow==9.4.0
opencv-python==4.7.0.72

